backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic-settings>=2.4.0
pandas>=2.2.0
numpy>=2.0.0
optuna>=3.6.0
python-multipart>=0.0.9
sqlalchemy>=2.0.32
alembic>=1.13.2
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.2.0
